import_export can force fields to be non-negative
If the atm -> ice scalar map is not positive definite, then ensure Qa and rhoa
are non-negative by setting all negative values to 0. Similarly, if the atm ->
ice flux map is not positive definite, ensure swvdr, swidr, swvdf, swidf,
frain, and fsnow are all non-negative.

The implementation is a little kludgy: component models in CESM do not have
access to public data in the coupler (otherwise this introduces circular
dependencies), so we can't just query the driver to find out about the map.
Instead, I introduce two new namelist variables in &setup_nml and buildnml sets
them based on the ATM2OCN map names in env_run.xml. Currently, the logic is "if
'patc' shows up in the file name, it's a patch map and therefore not positive
definite. Otherwise it's a positive definite map."