Drop skip_default_entry from atm2ice_ variables
Instead, default values for atm2ice_fmap_is_pos_def (and smap equivalent) are
both .true., but if fmaptype (or smaptype) is "patch" they are toggled to
false. The buildnml script sets config['fmaptype'] and config['smaptype'] based
on the str.find() command, though if other maps that are not positive definite
are introduced perhaps these if statements should be replaced with a function?